backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
websockets==14.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
numpy==2.2.1
